[s3c2416] Move INTC to its own source files
